Update design of rain gauge 3d print files
Changed design files for the modified and enlarged rain gauge:

Updated the funnel to enlarge and strengthen the base and support strut.

Widened the housing and modified re changes to the funnel mounting.

Added a new component to insert into the funnel exit hole to gradually
narrow the bore to more accurately guide the water over the bucket.

Widened and strengthened the housing floor.

Modified the design of the bucket to accept a larger magnet and thicker
axle.

Redesigned the reed switch holder of the bucket bracket and increased
the size of the axle holes to allow for a thicker axle.

Deleted the axle shaft caps as they are no longer required.

Added a cap to enable the magnet to be totally enclosed.